refactor: move CLI to interfaces/ directory
## Changes
- Create thetable/interfaces/ directory
- Move CLI logic from __main__.py to interfaces/cli.py
- Simplify __main__.py to thin entry point (5 lines)
- Update pyproject.toml script entry point
- Update test imports to thetable.interfaces.cli

## Structure
```
thetable/
├── __main__.py              # Thin entry point
├── interfaces/              # User interface layer
│   ├── __init__.py
│   └── cli.py              # CLI implementation
```

## Benefits
- Better separation of concerns
- CLI is one of many possible interfaces
- Easier to add web/API interfaces in the future
- Clearer package architecture

## Test Results
- All 36 tests passing ✅
