+++ thetable/interfaces/__init__.py
@@ -1,0 +1,1 @@
+"""User interfaces for TheTable"""
